LCAM-842: Moved the maat api related  schemas to maat-api folder and refactored the code
