+++ crown-court-contribution/src/main/java/uk/gov/justice/laa/crime/contribution/builder/AppealContributionResponseMapper.java
@@ -2,7 +2,7 @@ package uk.gov.justice.laa.crime.contribution.builder;
 
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
-import uk.gov.justice.laa.crime.contribution.model.AppealContributionResponse;
+import uk.gov.justice.laa.crime.contribution.model.maat_api.AppealContributionResponse;
 import uk.gov.justice.laa.crime.contribution.model.Contribution;
 
 import java.time.LocalDate;
