+++ crown-court-contribution/src/main/java/uk/gov/justice/laa/crime/contribution/builder/ContributionSummaryMapper.java
@@ -3,7 +3,7 @@ package uk.gov.justice.laa.crime.contribution.builder;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
 import uk.gov.justice.laa.crime.contribution.dto.ContributionsSummaryDTO;
-import uk.gov.justice.laa.crime.contribution.model.ContributionSummary;
+import uk.gov.justice.laa.crime.contribution.model.common.ContributionSummary;
 
 import static uk.gov.justice.laa.crime.contribution.util.DateUtil.convertDateToDateTime;
