+++ crown-court-contribution/src/main/java/uk/gov/justice/laa/crime/contribution/builder/GetContributionAmountRequestMapper.java
@@ -3,8 +3,8 @@ package uk.gov.justice.laa.crime.contribution.builder;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
 import uk.gov.justice.laa.crime.contribution.dto.CalculateContributionDTO;
-import uk.gov.justice.laa.crime.contribution.model.AppealContributionRequest;
-import uk.gov.justice.laa.crime.contribution.model.GetContributionAmountRequest;
+import uk.gov.justice.laa.crime.contribution.model.maat_api.AppealContributionRequest;
+import uk.gov.justice.laa.crime.contribution.model.maat_api.GetContributionAmountRequest;
 import uk.gov.justice.laa.crime.contribution.staticdata.enums.AssessmentResult;
 
 @Component
