+++ crown-court-contribution/src/main/java/uk/gov/justice/laa/crime/contribution/service/ContributionRulesService.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import uk.gov.justice.laa.crime.contribution.dto.ContributionVariationDTO;
-import uk.gov.justice.laa.crime.contribution.model.ApiCrownCourtOutcome;
-import uk.gov.justice.laa.crime.contribution.model.ApiCrownCourtSummary;
+import uk.gov.justice.laa.crime.contribution.model.maat_api.ApiCrownCourtOutcome;
+import uk.gov.justice.laa.crime.contribution.model.maat_api.ApiCrownCourtSummary;
 import uk.gov.justice.laa.crime.contribution.staticdata.entity.ContributionRulesEntity;
 import uk.gov.justice.laa.crime.contribution.staticdata.enums.CaseType;
 import uk.gov.justice.laa.crime.contribution.staticdata.enums.CrownCourtOutcome;
