+++ crown-court-contribution/src/main/java/uk/gov/justice/laa/crime/contribution/validation/CalculateContributionValidator.java
@@ -3,7 +3,7 @@ package uk.gov.justice.laa.crime.contribution.validation;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import uk.gov.justice.laa.crime.contribution.exeption.ValidationException;
-import uk.gov.justice.laa.crime.contribution.model.MaatCalculateContributionRequest;
+import uk.gov.justice.laa.crime.contribution.model.maat_api.MaatCalculateContributionRequest;
 import uk.gov.justice.laa.crime.contribution.staticdata.enums.AssessmentStatus;
 
 import java.time.LocalDateTime;
