+++ crown-court-contribution/src/main/resources/schemas/maat-api/MaatCalculateContributionRequest.json
@@ -148,7 +148,7 @@
       "description": "List of assessments performed",
       "minItems": 1,
       "items": {
-        "$ref": "common/Assessment.json"
+        "$ref": "../common/Assessment.json"
       }
     },
     "removeContributions": {
