+++ crown-court-contribution/src/main/resources/schemas/maat-api/apiCalculateHardshipByDetailRequest.json
