+++ crown-court-contribution/src/main/resources/schemas/maat-api/apiCalculateHardshipByDetailResponse.json
