+++ crown-court-contribution/src/test/java/uk/gov/justice/laa/crime/contribution/service/AppealContributionServiceTest.java
@@ -11,6 +11,8 @@ import uk.gov.justice.laa.crime.contribution.builder.GetContributionAmountReques
 import uk.gov.justice.laa.crime.contribution.data.builder.TestModelDataBuilder;
 import uk.gov.justice.laa.crime.contribution.dto.CalculateContributionDTO;
 import uk.gov.justice.laa.crime.contribution.model.*;
+import uk.gov.justice.laa.crime.contribution.model.common.Assessment;
+import uk.gov.justice.laa.crime.contribution.model.maat_api.*;
 import uk.gov.justice.laa.crime.contribution.staticdata.enums.AssessmentResult;
 import uk.gov.justice.laa.crime.contribution.staticdata.enums.CaseType;
