+++ crown-court-contribution/src/test/java/uk/gov/justice/laa/crime/contribution/validation/AppealContributionValidatorTest.java
@@ -3,7 +3,9 @@ package uk.gov.justice.laa.crime.contribution.validation;
 import org.junit.jupiter.api.Test;
 import uk.gov.justice.laa.crime.contribution.data.builder.TestModelDataBuilder;
 import uk.gov.justice.laa.crime.contribution.exeption.ValidationException;
-import uk.gov.justice.laa.crime.contribution.model.*;
+import uk.gov.justice.laa.crime.contribution.model.common.Assessment;
+import uk.gov.justice.laa.crime.contribution.model.maat_api.LastOutcome__1;
+import uk.gov.justice.laa.crime.contribution.model.maat_api.MaatCalculateContributionRequest;
 import uk.gov.justice.laa.crime.contribution.staticdata.enums.AssessmentStatus;
 
 import java.time.LocalDateTime;
